dart documentation has been updated
