+++ lib/doc_scan_kit.dart
@@ -1,4 +1,4 @@
-export 'package:doc_scan_kit/src/doc_scan.dart';
+export 'package:doc_scan_kit/src/doc_scan_kit.dart';
 export 'package:doc_scan_kit/src/models/doc_scan_kit_format.dart';
 export 'package:doc_scan_kit/src/models/doc_scan_kit_options_android.dart';
 export 'package:doc_scan_kit/src/models/doc_scan_kit_options_ios.dart';
